Pull out ffi bindings into sub-crate
