+++ Cargo.toml
@@ -4,11 +4,7 @@ version = "0.1.0"
 authors = ["Ethan Frey <ethanfrey@users.noreply.github.com>"]
 edition = "2018"
 
-[lib]
-crate-type = ["cdylib"]
-
 [dependencies]
-libc = "0.2"
 
-[build-dependencies]
-cbindgen = "0.6"
+[workspace]
+members = ["lib/*"]
