+++ rust-binding-investigation.iml
@@ -7,6 +7,16 @@
       <sourceFolder url="file://$MODULE_DIR$/examples" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/tests" isTestSource="true" />
       <sourceFolder url="file://$MODULE_DIR$/benches" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/lib/rust_ffi/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/lib/rust_ffi/examples" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/lib/rust_ffi/tests" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/lib/rust_ffi/benches" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/lib/rust_wasm/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/lib/rust_wasm/examples" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/lib/rust_wasm/tests" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/lib/rust_wasm/benches" isTestSource="true" />
+      <excludeFolder url="file://$MODULE_DIR$/lib/rust_ffi/target" />
+      <excludeFolder url="file://$MODULE_DIR$/lib/rust_wasm/target" />
       <excludeFolder url="file://$MODULE_DIR$/target" />
     </content>
     <orderEntry type="inheritedJdk" />
